+++ tracy/go.mod
@@ -1,3 +1,3 @@
-module github.com/decentplatforms/appkit/tracy
+module github.com/decentplatforms/appkit/logf
 
 go 1.21.1
